DOMCore/tests/submissions/Ms2ger/EventTarget-dispatchEvent.html

<!DOCTYPE html>
<title>EventTarget.dispatchEvent</title>
<link rel="author" title="Olli Pettay" href="mailto:Olli.Pettay@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var e = document.createEvent("Event")
  assert_equals(e.type, "", "Event type should be empty string before initialization")
  assert_throws("INVALID_STATE_ERR", function() { document.dispatchEvent(e) })
})
</script>
